feat(jetbrains): add OpenApiSpecNormalizer with unit tests and decouple logic from Gradle task
Introduce OpenApiSpecNormalizer as a pure-function object that owns all
JSON parsing, tag deduplication, and path rewriting logic. The Gradle
task now delegates to this normalizer, keeping task wiring minimal.

Add JUnit 5 test infrastructure to the build-tasks module and cover the
normalizer with tests for successful renames, no-op passthrough, and
missing-operation error handling. Replace Array usages in config data
classes with List for idiomatic Kotlin collections.
