+++ packages/kilo-jetbrains/build-tasks/build.gradle.kts
@@ -8,6 +8,11 @@ repositories {
 
 dependencies {
     implementation(libs.kotlinx.serialization.json)
+    testImplementation(kotlin("test"))
+}
+
+tasks.test {
+    useJUnitPlatform()
 }
 
 gradlePlugin {
