+++ packages/kilo-jetbrains/build-tasks/src/main/kotlin/normalization/NormalizeOpenApiSpecConfig.kt
@@ -2,22 +2,22 @@ package normalization
 
 internal data class DuplicateTagRule(
     val original: String,
-    val dedups: Array<TagDedup>,
+    val dedups: List<TagDedup>,
 )
 
 internal data class TagDedup(
     val name: String,
-    val ops: Array<String> = emptyArray(),
+    val ops: List<String> = emptyList(),
 )
 
-internal val duplicateTagRules = arrayOf(
+internal val duplicateTagRules = listOf(
     DuplicateTagRule(
         original = "pty",
-        dedups = arrayOf(
+        dedups = listOf(
             TagDedup(name = "pty"),
             TagDedup(
                 name = "pty-connect",
-                ops = arrayOf("pty.connect"),
+                ops = listOf("pty.connect"),
             ),
         ),
     ),
